+++ examples/resources/iosxe_interface_tunnel/resource.tf
@@ -45,4 +45,5 @@ resource "iosxe_interface_tunnel" "example" {
   load_interval                    = 30
   snmp_trap_link_status            = false
   logging_event_link_status_enable = true
+  tunnel_vrf                       = "VRF1"
 }