+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.9 (unreleased)
 
 - Fix idempotency issue with `secret` attribute of `iosxe_username` resource
+- Add `tunnel_vrf` attribute to `iosxe_interface_tunnel` resource and data source
 
 ## 0.5.8